+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Interface
 
+- [Hide blank side menu](https://github.com/catalpainternational/myeqip-portal/pull/95)
+- [WebPage and Blocks have help texts](https://github.com/catalpainternational/myeqip-portal/pull/99)
+- [Ministry logo is a link](https://github.com/catalpainternational/myeqip-portal/pull/101)
 - [Line block](https://github.com/catalpainternational/myeqip-portal/pull/103)
-
-### Code
 
 ## 0.0.3